backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.12.5
pydantic-settings==2.12.0

# File uploads
python-multipart==0.0.21

# Environment variables
python-dotenv==1.2.1

# PDF parsing
pypdf2==3.0.1

# AI (Claude)
anthropic==0.7.8

# Agentic workflows
agno==2.3.21

# AWS
boto3==1.42.21

# RAG - Vector Store & Embeddings (all free, runs locally)
chromadb==0.4.22
sentence-transformers==2.2.2

# Groq LLM
groq==1.0.0